services/pricingservice/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
grpcio==1.46.3
    # via -r requirements.in
protobuf==4.21.1
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
six==1.16.0
    # via grpcio
